requirements.txt
langchain
huggingface
transformers
torch
streamlit
streamlit-chat
python-dotenv
openai